clang build scripts and changes spotted - clang wont build OOCore yet - crashes clang!

+++ src/OOBase/Atomic.h
@@ -47,7 +47,7 @@
 #elif defined(HAVE___SYNC_TEST_AND_SET)
 
 /* Define to if you have atomic exchange for 32bit values */
-#define ATOMIC_EXCH_32(t,v) __sync_lock_test_and_set((long volatile*)(t),v)
+#define ATOMIC_EXCH_32(t,v) __sync_lock_test_and_set((int volatile*)(t),v)
 
 /* Define to if you have atomic exchange for 64bit values */
 #define ATOMIC_EXCH_64(t,v) __sync_lock_test_and_set((long long volatile*)(t),v)
@@ -55,10 +55,10 @@
 #if defined(HAVE___SYNC_ADD_AND_FETCH)
 
 /* Define to if you have atomic inc and dec for 32bit values */
-#define ATOMIC_INC_32(t) __sync_add_and_fetch((long*)(t),1)
-#define ATOMIC_DEC_32(t) __sync_sub_and_fetch((long*)(t),1)
-#define ATOMIC_ADD_32(t,v) __sync_add_and_fetch((long volatile*)(t),v)
-#define ATOMIC_SUB_32(t,v) __sync_sub_and_fetch((long volatile*)(t),v)
+#define ATOMIC_INC_32(t) __sync_add_and_fetch((int*)(t),1)
+#define ATOMIC_DEC_32(t) __sync_sub_and_fetch((int*)(t),1)
+#define ATOMIC_ADD_32(t,v) __sync_add_and_fetch((int volatile*)(t),v)
+#define ATOMIC_SUB_32(t,v) __sync_sub_and_fetch((int volatile*)(t),v)
 
 /* Define to if you have atomic exchange for 64bit values */
 #define ATOMIC_INC_64(t) __sync_add_and_fetch((long long volatile*)(t),1)
